Move BasketItem and OrderForm to own classes
